Crate libp2p_tcp

Implementation of the libp2p Transport trait for TCP/IP.

Uses the tokio library.

Usage

Example:

extern crate libp2p_tcp;
use libp2p_tcp::TcpConfig;

let tcp = TcpConfig::new();

The TcpConfig structs implements the Transport trait of the swarm library. See the documentation of swarm and of libp2p in general to learn how to use the Transport trait.

Structs

TcpConfig
Represents the configuration for a TCP/IP transport capability for libp2p.

TcpDialFut
Future that dials a TCP/IP address.

TcpListenStream
Stream that listens on an TCP/IP address.

TcpListener
Stream of ListenerEvents.

TcpTransStream
Wraps around a TcpStream and adds logging for important events.