Struct libp2p_tcp::TcpConfig

pub struct TcpConfig { /* fields omitted */ }

Represents the configuration for a TCP/IP transport capability for libp2p.

The TCP sockets created by libp2p will need to be progressed by running the futures and streams obtained by libp2p through the tokio reactor.

Methods

impl TcpConfig

pub fn new() -> TcpConfig

Creates a new configuration object for TCP/IP.

pub fn recv_buffer_size(self, value: usize) -> Self

Sets the size of the recv buffer size to set for opened sockets.

pub fn send_buffer_size(self, value: usize) -> Self

Sets the size of the send buffer size to set for opened sockets.

pub fn ttl(self, value: u32) -> Self

Sets the TTL to set for opened sockets.

pub fn keepalive(self, value: Option<Duration>) -> Self

Sets the keep alive pinging duration to set for opened sockets.

pub fn nodelay(self, value: bool) -> Self

Sets the TCP_NODELAY to set for opened sockets.