Crate libp2p_tcp

Implementation of the libp2p Transport trait for TCP/IP.

Usage

This crate provides two structs, TcpConfig and TokioTcpConfig, depending on which features are enabled.

Both the TcpConfig and TokioTcpConfig structs implement the Transport trait of the core library. See the documentation of core and of libp2p in general to learn how to use the Transport trait.

Structs

TcpConfig
Represents the configuration for a TCP/IP transport capability for libp2p.

TcpListenStream
Stream that listens on an TCP/IP address.

TcpTransStream
Wraps around a TcpStream and adds logging for important events.