Crate libp2p_tcp

Implementation of the libp2p libp2p_core::Transport trait for TCP/IP.

Usage

This crate provides a [async_io::Transport] and tokio::Transport, depending on the enabled features, which implement the libp2p_core::Transport trait for use as a transport with libp2p-core or libp2p-swarm.

Modules

tokio

Structs

Config

The configuration for a TCP/IP transport capability for libp2p.

Transport

An abstract libp2p_core::Transport implementation.